# Database cachednode callback walk