Replaces https://github.com/ethereum/go-ethereum/pull/20527 , this one does not contain all the other experimental fluff. 

In the trie database, there are several operations that recursively walk children of nodes, and collect them into lists of hashes. ~~This means that on the leaf level, all the upper levels hold N lists with hashes. So at root level, the list contains all leaf hashes, at level below, a list with 1/16th of the entries, and so on.~~

This PR changes it to walk the nodes with a callback instead. Needs to be benchmarked and tested IRL, of course. 


before
```
goos: linux
goarch: amd64
pkg: github.com/ethereum/go-ethereum/trie
BenchmarkDerefRootFixedSize/10-6                   61821             19229 ns/op           14609 B/op         38 allocs/op
BenchmarkDerefRootFixedSize/100-6                  20331             58074 ns/op           67856 B/op        142 allocs/op
BenchmarkDerefRootFixedSize/1K-6                    1717            636877 ns/op          696097 B/op       1369 allocs/op
BenchmarkDerefRootFixedSize/10K-6                    145           8177885 ns/op         7116049 B/op      13908 allocs/op
BenchmarkDerefRootFixedSize/100K-6                    13          85116617 ns/op        70853392 B/op     138395 allocs/op
```
after
```
goos: linux
goarch: amd64
pkg: github.com/ethereum/go-ethereum/trie
BenchmarkDerefRootFixedSize/10-6                   76370             16321 ns/op             785 B/op         11 allocs/op
BenchmarkDerefRootFixedSize/100-6                  24958             47222 ns/op             784 B/op         11 allocs/op
BenchmarkDerefRootFixedSize/1K-6                    2838            419798 ns/op             784 B/op         11 allocs/op
BenchmarkDerefRootFixedSize/10K-6                    202           5425507 ns/op             784 B/op         11 allocs/op
BenchmarkDerefRootFixedSize/100K-6                    20          55343687 ns/op             784 B/op         11 allocs/op
```