@karalabe 2020-01-09
It only gathers the direct children of a node, not recursively the entire subtrie.