@holiman 2020-01-09
> It only gathers the direct children of a node, not recursively the entire subtrie.

You are correct. I suspect the problem is that we did a `children := make([]common.Hash, 0, 16)` at every junction, even though many don't even have children, let alone 16.  So it causes a lot of tiny mallocs that were not even used.

https://github.com/ethereum/go-ethereum/pull/20529/files#diff-e49891341ce9d49a244d9ead8df3d072L186