@holiman 2020-01-09
Apparently this code doesn't work, though. It changes the semantics, as it exceutes the callback immediately, instead of collecting the children first, so it breaks things.