@holiman 2020-01-09
Giving this a shot on the benchmarkers, replacing the current run (https://geth-bench.ethdevops.io/d/Jpk-Be5Wk/dual-geth?orgId=1&var-exp=mon08&var-master=mon09&var-percentile=50&from=1578482122263&to=now )  without restarting them. The experimental (`mon08`) was at `5.021M` when it was replaced.