@holiman 2020-01-10
This actually does seem to make a difference. After running for a couple of hours, I checked the same section of blocks for both of these. 
First is the `mon09` (`master) between `5.1M - 5.2M`:

![Screenshot_2020-01-10 Dual Geth - Grafana](https://user-images.githubusercontent.com/142290/72134024-2d2ede00-3383-11ea-96c4-c5fc13e71e90.png)

Although absolute time-comparisons between machines is a bit flakey, we can see that `commit` averages `13.16ms`, above `account commit` at `11.62ms`.

For this pr, `mon08`, this is the corresponding chart: 
![Screenshot_2020-01-10 Dual Geth - Grafana(1)](https://user-images.githubusercontent.com/142290/72134138-7bdc7800-3383-11ea-9bba-be996a0d2bb2.png)

In this chart, `account commit` is at `11.37ms` (which is _very_ close, even in absolute numbers, to `11.62ms`), whereas `commit` has dropped to third place at `11.13ms`.